@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <LanguageVersion>8.0</LanguageVersion>
+    <TargetFramework>net9.0</TargetFramework>
+    <LanguageVersion>latest</LanguageVersion>
     <Nullable>enable</Nullable>
 
     <PackAsTool>true</PackAsTool>
@@ -15,10 +15,10 @@
     <Copyright>Copyright Mike Goatly</Copyright>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageProjectUrl>https://github.com/mikegoatly/html2md</PackageProjectUrl>
-    <Version>2.1.9</Version>
+    <Version>3.0.0</Version>
     <PackageTags>convert-html convert-markdown html markdown conversion</PackageTags>
     <RepositoryUrl>https://github.com/mikegoatly/html2md</RepositoryUrl>
-    <PackageReleaseNotes>Added support for extracting Front Matter metadata</PackageReleaseNotes>
+    <PackageReleaseNotes>Updated to .NET 9</PackageReleaseNotes>
   </PropertyGroup>
 
   <ItemGroup>
